requirements.txt
numpy>=1.18
starlink-pyast>=3.14.4
astropy
